backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic[email]
edge-tts
piper-tts
PyJWT
bcrypt
cryptography
